(77) Rename SimpleMDE into CodeMirror
