+++ src/common/HaskellDo/CodeMirror/Types.hs
@@ -13,16 +13,12 @@
  - See the License for the specific language governing permissions and
  - limitations under the License.
  -}
-module HaskellDo.SimpleMDE.State where
+module HaskellDo.CodeMirror.Types where
 
-import Transient.Move
-import HaskellDo.SimpleMDE.Types
+data Action
+    = NewContent String
+    deriving (Read, Show)
 
-initialState :: State
-initialState = State
-    { content = ""
-    }
-
-update :: Action -> State -> Cloud State
-update (NewContent newContent) state = do
-    return (state { content = newContent } )
+data State = State
+    { content :: String
+    } deriving (Read, Show)
